code cleanup

+++ urgap/uinit.py
@@ -21,10 +21,12 @@ def copy_resources_if_needed(
     target_resources_path.mkdir(exist_ok=True)
     for rfile in source_resources_path.glob("**/*"):
         if rfile.is_file() is True:
+            if rfile.name.startswith("."):
                 continue
             target_rfile = Path(
                 str(rfile).replace(
             )
+            if target_rfile.exists() is False or force is True:
                 target_rfile.parent.mkdir(exist_ok=True, parents=True)
                 shutil.copy(rfile, target_rfile)
                 if force is False:
